requirements.txt
Flask==2.0.2
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.5.1
pandas==1.3.4
matplotlib==3.4.3